AABB2 rendering and closest point working
[git-p4: depot-paths = "//depot/C35/Students/cwsullivan/SD/": change = 437742]

+++ MathVisualTests/Code/Game/TestShapeTriangle.cpp
@@ -21,7 +21,7 @@ void TestShapeTriangle::Render() const
 {
 	std::vector<Vertex> closePoint;
 	AddVertsForDisc2D( closePoint, Vec2( 0.f, 0.f ), 1.0f, Rgba8( 255, 215, 0, 255 ) );
-	TransformVertexArrayXY3D( closePoint.size(), closePoint.data(), .5f, 0.f, m_closestPoint );
+	TransformVertexArrayXY3D( static_cast< int >( closePoint.size() ), closePoint.data(), .5f, 0.f, m_closestPoint );
 	g_engine->m_render->DrawVertexArray( ( int )m_lineVerts.size(), m_lineVerts.data() );
 	g_engine->m_render->DrawVertexArray( closePoint );
 }
